+++ Automata.Engine/Rendering/OpenGL/BufferAllocator.cs
@@ -16,7 +16,7 @@ namespace Automata.Engine.Rendering.OpenGL
 
         public int AllocatedBuffers => _NativeMemoryPool.RentedBlocks;
 
-        public unsafe BufferAllocator(GL gl, uint size) : base(gl)
+        public unsafe BufferAllocator(GL gl, nuint size) : base(gl)
         {
             Handle = GL.CreateBuffer();
             GL.NamedBufferStorage(Handle, size, (void*)null!, (uint)_STORAGE_MASK);
@@ -24,7 +24,7 @@ namespace Automata.Engine.Rendering.OpenGL
             _NativeMemoryPool = new NativeMemoryPool(pointer, size);
         }
 
-        public IMemoryOwner<T> Rent<T>(uint size) where T : unmanaged => _NativeMemoryPool.Rent<T>(size);
+        public IMemoryOwner<T> Rent<T>(int size) where T : unmanaged => _NativeMemoryPool.Rent<T>(size);
 
         public void Bind(BufferTargetARB target) => GL.BindBuffer(target, Handle);
